Truly random result
Updated solution with a truly random result that allows for any possibility. Uses two lists instead of one. Random names are chosen from giver, receiver and then checked to make sure they are not the same and there is no swapping. This version does not use recursion. This version is slower.